Live排版优化

+++ app/src/main/java/com/onecric/live/model/LiveBean.java
@@ -116,7 +116,7 @@ public class LiveBean {
         this.match_id = match_id;
     }
 
-    public int getStarttime() {
+    public long getStarttime() {
         return starttime;
     }
